requirements.txt
pandas
sqlalchemy
python-dotenv
tabula-py
jpype1
requests
beautifulsoup4
pytest